+++ processor/resourcedetectionprocessor/internal/gcp/gcp_test.go
@@ -139,10 +139,32 @@ func TestDetect(t *testing.T) {
 			}),
 		},
 		{
-			desc: "App Engine",
+			desc: "App Engine Standard",
 			detector: &detector{detector: &fakeGCPDetector{
 				projectID:                 "my-project",
-				cloudPlatform:             gcp.AppEngine,
+				cloudPlatform:             gcp.AppEngineStandard,
+				appEngineServiceInstance:  "1472385723456792345",
+				appEngineAvailabilityZone: "us-central1-c",
+				appEngineRegion:           "us-central1",
+				appEngineServiceName:      "my-service",
+				appEngineServiceVersion:   "123456",
+			}},
+			expectedResource: internal.NewResource(map[string]interface{}{
+				conventions.AttributeCloudProvider:         conventions.AttributeCloudProviderGCP,
+				conventions.AttributeCloudAccountID:        "my-project",
+				conventions.AttributeCloudPlatform:         conventions.AttributeCloudPlatformGCPAppEngine,
+				conventions.AttributeCloudRegion:           "us-central1",
+				conventions.AttributeCloudAvailabilityZone: "us-central1-c",
+				conventions.AttributeFaaSName:              "my-service",
+				conventions.AttributeFaaSVersion:           "123456",
+				conventions.AttributeFaaSID:                "1472385723456792345",
+			}),
+		},
+		{
+			desc: "App Engine Flex",
+			detector: &detector{detector: &fakeGCPDetector{
+				projectID:                 "my-project",
+				cloudPlatform:             gcp.AppEngineFlex,
 				appEngineServiceInstance:  "1472385723456792345",
 				appEngineAvailabilityZone: "us-central1-c",
 				appEngineRegion:           "us-central1",
@@ -285,11 +307,25 @@ func (f *fakeGCPDetector) FaaSCloudRegion() (string, error) {
 	return f.faaSCloudRegion, nil
 }
 
-func (f *fakeGCPDetector) AppEngineAvailabilityZoneAndRegion() (string, string, error) {
+func (f *fakeGCPDetector) AppEngineFlexAvailabilityZoneAndRegion() (string, string, error) {
 	if f.err != nil {
 		return "", "", f.err
 	}
 	return f.appEngineAvailabilityZone, f.appEngineRegion, nil
+}
+
+func (f *fakeGCPDetector) AppEngineStandardAvailabilityZone() (string, error) {
+	if f.err != nil {
+		return "", f.err
+	}
+	return f.appEngineAvailabilityZone, nil
+}
+
+func (f *fakeGCPDetector) AppEngineStandardCloudRegion() (string, error) {
+	if f.err != nil {
+		return "", f.err
+	}
+	return f.appEngineRegion, nil
 }
 
 func (f *fakeGCPDetector) AppEngineServiceName() (string, error) {